requirements.txt
nextcord==2.0.0a8
python-dotenv==0.19.2
pymongo==4.0.1
dnspython==2.2.0
InfixParser==3.1.4
requests==2.27.1
humanfriendly==10.0
NextcordUtils==1.0